BTA-1011 Update API documentation for account validation
Updates the API documentation to reflect that in the future account validation and name enquiry might be supported across multiple currencies, countries and potentially even on mobile. Also renames the `account_id` field to `bank_account` to make it more consistent with how we address bank details when creating transactions.

Note that calling the endpoint the old way, e.g. skipping `currency` and `method` and using `account_id` instead of `bank_account` is still fully supported by the system